refactor: move utils into common superclass
